HW14 is completed (#22)

+++ hw12_13_14_15_calendar/models/notification.go
@@ -8,7 +8,7 @@ import (
 
 type Notification struct {
 	EventID  uuid.UUID
-	UserID   int32
+	UserID   int64
 	Title    string
 	Datetime time.Time
 }
